+++ package/example/lib/advanced_example/screens/custom_field/widgets/color_controller.dart
@@ -1,5 +1,5 @@
-import 'package:easy_form_kit/easy_form_kit.dart';
 import 'package:flutter/material.dart';
+import 'package:easy_form_kit/easy_form_kit.dart';
 
 class ColorController extends EasyFormFieldController<Color> {
   ColorController(Color value) : super(value);