+++ package/example/lib/advanced_example/screens/custom_field/widgets/color_form_field.dart
@@ -1,7 +1,7 @@
-import 'package:easy_form_example/screens/custom_field/widgets/color_controller.dart';
-import 'package:easy_form_example/screens/custom_field/widgets/color_field.dart';
-import 'package:easy_form_kit/easy_form_kit.dart';
 import 'package:flutter/material.dart';
+import 'package:easy_form_kit/easy_form_kit.dart';
+import 'package:easy_form_example/advanced_example/screens/custom_field/widgets/color_controller.dart';
+import 'package:easy_form_example/advanced_example/screens/custom_field/widgets/color_field.dart';
 
 class ColorFormField extends EasyCustomFormField<Color, ColorController> {
   ColorFormField({